requirements.txt
openai
pandas
tqdm
pyyaml
tenacity
python-dotenv